+++ CHANGELOG.md
@@ -14,6 +14,10 @@ Bugsnag Notifiers on other platforms.
 * BugsnagCrashReport is now BugsnagEvent
   [#449](https://github.com/bugsnag/bugsnag-cocoa/pull/449)
 
+* Add a configuration option to filter breadcrumbs by type. Use
+  `config.enabledBreadcrumbTypes` to enable or disable particular types of
+  breadcrumbs from being captured.
+
 * Added a designated initializer to `BugsnagConfiguration` and removed functionality
   from the default convenience `init()` to ensure that `apiKey` has a value set.  The `apiKey`
   must now be a correctly formatted one to be accepted.